+++ QModel/src/models/q_multi_model.py
@@ -307,8 +307,7 @@ class QMultiModel:
         Example:
             best_hyperparams = self.tune(evaluations=300)
         """
-        print(f"[STATUS] Running model tuning for {
-              evaluations} max iterations")
+        print(f"[STATUS] Running model tuning for {evaluations} max iterations")
         space = {
             "max_depth": hp.choice("max_depth", np.arange(1, 20, 1, dtype=int)),
             "eta": hp.uniform("eta", 0, 1),
@@ -414,8 +413,7 @@ class QPredictor:
             relative_root = os.path.join(Architecture.get_path(), "QATCH")
         else:
             relative_root = os.getcwd()
-        pickle_path = os.path.join(
-            relative_root, "QModel/SavedModels/label_{}.pkl")
+        pickle_path = os.path.join(relative_root, "QModel/SavedModels/label_{}.pkl")
         for i in range(3):
             with open(pickle_path.format(i), "rb") as file:
                 setattr(self, f"__label_{i}__", pickle.load(file))
@@ -479,8 +477,7 @@ class QPredictor:
             return adj_prediction, (lq_idx, uq_idx)
 
         lq_idx = next((i for i, x in enumerate(adj) if x == 1), -1) + i
-        uq_idx = next((i for i, x in reversed(
-            list(enumerate(adj))) if x == 1), -1) + i
+        uq_idx = next((i for i, x in reversed(list(enumerate(adj))) if x == 1), -1) + i
         return prediction, (lq_idx, uq_idx)
 
     def find_and_sort_peaks(self, signal):
@@ -511,7 +508,7 @@ class QPredictor:
         slope_change = []
         for i in range(len(diff) - window_size + 1):
             # Calculate the slope change over the current window
-            window_slope_change = np.mean(np.diff(diff[i: i + window_size]))
+            window_slope_change = np.mean(np.diff(diff[i : i + window_size]))
             slope_change.append(window_slope_change)
 
         slope_change = np.array(slope_change)
@@ -574,8 +571,7 @@ class QPredictor:
 
     def adjustment_poi_1(self, guess, diss_raw):
 
-        zero_slope = self.find_zero_slope_regions(
-            self.normalize(diss_raw), 0.0075, 100)
+        zero_slope = self.find_zero_slope_regions(self.normalize(diss_raw), 0.0075, 100)
         adjusted_guess = guess
 
         if len(zero_slope) >= 2:
@@ -699,7 +695,10 @@ class QPredictor:
         diff_points = np.array(diff_peaks)
         x_min, x_max = bounds
         candidates = np.append(candidates, guess)
-        candidate_density = len(candidates) / (x_max - x_min)
+        if x_max - x_min > 0:
+            candidate_density = len(candidates) / (x_max - x_min)
+        else:
+            candidate_density = 1
         if candidate_density < 0.02 or (guess < x_min or guess > x_max):
             # Filter RF points within the bounds
             rf_points = np.concatenate((rf_points, diss_points))
@@ -744,8 +743,7 @@ class QPredictor:
             closest_rf_point = filtered_rf_points[closest_rf_idx]
 
             # Calculate distances from candidate points to the closest RF point
-            distances_to_closest_rf = np.abs(
-                candidate_points - closest_rf_point)
+            distances_to_closest_rf = np.abs(candidate_points - closest_rf_point)
 
             # Find the closest candidate point to the closest RF point
             closest_candidate_idx = np.argmin(distances_to_closest_rf)
@@ -798,7 +796,10 @@ class QPredictor:
         diff_points = np.array(diff_peaks)
         np.concatenate((rf_points, diff_points, diss_points))
         x_min, x_max = bounds
-        candidate_density = len(candidates) / (x_max - x_min)
+        if x_max - x_min:
+            candidate_density = len(candidates) / (x_max - x_min)
+        else:
+            candidate_density = 1
         if candidate_density < 0.01:
             zero_slope = self.find_zero_slope_regions(rf)
             # Filter RF points within the bounds
@@ -841,8 +842,7 @@ class QPredictor:
             closest_rf_point = filtered_rf_points[closest_rf_idx]
 
             # Calculate distances from candidate points to the closest RF point
-            distances_to_closest_rf = np.abs(
-                candidate_points - closest_rf_point)
+            distances_to_closest_rf = np.abs(candidate_points - closest_rf_point)
 
             # Find the closest candidate point to the closest RF point
             closest_candidate_idx = np.argmin(distances_to_closest_rf)
@@ -910,8 +910,7 @@ class QPredictor:
         columns_to_drop = ["Date", "Time", "Ambient", "Temperature"]
         if not all(col in df.columns for col in columns_to_drop):
             raise ValueError(
-                f"[QModelPredict predict()]: Input data must contain the following columns: {
-                    columns_to_drop}"
+                f"[QModelPredict predict()]: Input data must contain the following columns: {columns_to_drop}"
             )
 
         df = df.drop(columns=columns_to_drop)
@@ -979,8 +978,12 @@ class QPredictor:
 
         # Process data using QDataPipeline
         qdp = QDataPipeline(file_buffer_2)
-        diss_raw = qdp.__dataframe__["Dissipation"]
-        rel_time = qdp.__dataframe__["Relative_time"]
+        qdp2 = QDataPipeline(file_buffer_2)
+        # t_delta = qdp2.find_time_delta()
+        # if t_delta > 0:
+        #     qdp2.downsample(k=t_delta,factor=20)
+        diss_raw = qdp2.__dataframe__["Dissipation"]
+        rel_time = qdp2.__dataframe__["Relative_time"]
         qdp.preprocess(poi_filepath=None)
         diff_raw = qdp.__difference_raw__
         df = qdp.get_dataframe()
@@ -1072,18 +1075,18 @@ class QPredictor:
             poi_1_guess=poi_1,
         )
         # poi_2 = emp_points[1]
-        poi_4 = self.adjustmet_poi_4(
-            df, candidates_4, extracted_4, act[3], bounds_4)
+        poi_4 = self.adjustmet_poi_4(df, candidates_4, extracted_4, act[3], bounds_4)
 
         # Hot fix to prevent out of order poi_4 and poi_5
-        if bounds_5[0] < poi_4:
-            lst = list(bounds_5)
-            lst[0] = poi_4 + 1
-            bounds_5 = tuple(lst)
+        # if bounds_5[0] < poi_4:
+        #     lst = list(bounds_5)
+        #     lst[0] = poi_4 + 1
+        #     bounds_5 = tuple(lst)
 
         poi_5 = self.adjustmet_poi_5(
             df, candidates_5, extracted_5, start_5, act[4], bounds_5
         )
+        # poi_5 = np.argmax(adj_5)
         if poi_1 >= poi_2:
             poi_1 = adj_1
         poi_3 = np.argmax(adj_3)
@@ -1104,7 +1107,7 @@ class QPredictor:
         def sort_and_remove_point(arr, point):
             arr = np.array(arr)
             if len(arr) > MAX_GUESSES - 1:
-                arr = arr[:MAX_GUESSES - 1]
+                arr = arr[: MAX_GUESSES - 1]
             arr.sort()
             return arr[arr != point]
 
@@ -1128,6 +1131,12 @@ class QPredictor:
         confidence_4 = np.array(adj_4)[candidates_4]
         confidence_5 = np.array(adj_5)[candidates_5]
         confidence_6 = np.array(adj_6)[candidates_6]
+        # confidence_1 = []
+        # confidence_2 = []
+        # confidence_3 = []
+        # confidence_4 = []
+        # confidence_5 = []
+        # confidence_6 = []
 
         # TODO: Adjust 1st confidence to be better than 2nd guess